refactor(blocks): move block structs and parsing logic
